+++ app/src/main/java/com/depromeet/muyaho/ui/addstock/AddStockActivity.kt
@@ -1,4 +1,4 @@
-package com.depromeet.muyaho.addstock
+package com.depromeet.muyaho.ui.addstock
 
 import androidx.activity.viewModels
 import com.depromeet.muyaho.R
