+++ app/src/main/java/com/depromeet/muyaho/ui/board/BoardFragment.kt
@@ -1,4 +1,4 @@
-package com.depromeet.muyaho.calc
+package com.depromeet.muyaho.ui.board
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -14,10 +14,10 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [CalcFragment.newInstance] factory method to
+ * Use the [BoardFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class CalcFragment : Fragment() {
+class BoardFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -35,7 +35,7 @@ class CalcFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_calc, container, false)
+        return inflater.inflate(R.layout.fragment_board, container, false)
     }
 
     companion object {
@@ -45,12 +45,12 @@ class CalcFragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment CalcFragment.
+         * @return A new instance of fragment BoardFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            CalcFragment().apply {
+            BoardFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
