+++ app/src/main/java/com/depromeet/muyaho/ui/calc/CalcFragment.kt
@@ -1,4 +1,4 @@
-package com.depromeet.muyaho.my
+package com.depromeet.muyaho.ui.calc
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -14,10 +14,10 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [MyFragment.newInstance] factory method to
+ * Use the [CalcFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class MyFragment : Fragment() {
+class CalcFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -35,7 +35,7 @@ class MyFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_my, container, false)
+        return inflater.inflate(R.layout.fragment_calc, container, false)
     }
 
     companion object {
@@ -45,12 +45,12 @@ class MyFragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment MyFragment.
+         * @return A new instance of fragment CalcFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            MyFragment().apply {
+            CalcFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
