+++ app/src/main/res/layout/fragment_add_stock_select_type.xml
@@ -5,7 +5,7 @@
     <data>
         <variable
             name="vm"
-            type="com.depromeet.muyaho.addstock.AddStockSelectTypeViewModel" />
+            type="com.depromeet.muyaho.ui.addstock.AddStockSelectTypeViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
