+++ app/src/main/res/layout/fragment_board.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".board.BoardFragment">
+    tools:context=".ui.board.BoardFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
