+++ app/src/main/res/layout/fragment_calc.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".calc.CalcFragment">
+    tools:context=".ui.calc.CalcFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
