+++ app/src/main/res/layout/fragment_home.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="vm"
-            type="com.depromeet.muyaho.home.HomeViewModel" />
+            type="com.depromeet.muyaho.ui.home.HomeViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
