+++ app/src/main/res/navigation/nav_my.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/fragment_my"
-        android:name="com.depromeet.muyaho.my.MyFragment"
+        android:name="com.depromeet.muyaho.ui.my.MyFragment"
         tools:layout="@layout/fragment_my">
 
     </fragment>
